refactor: split 6 oversized modules into submodules
Split 6 source files that exceeded 400 LOC into directory-based
submodules, each file now under 400 LOC:

- kuzu_store.rs (1236 LOC) → kuzu_store/{mod,connection,schema,helpers,queries,tests}.rs
- kuzu_backend.rs (948 LOC) → kuzu_backend/{mod,storage,search,tests}.rs
- ladybug_backend.rs (840 LOC) → ladybug_backend/{mod,storage,search,tests}.rs
- pattern_recognition.rs (706 LOC) → pattern_recognition/{mod,detection,scoring,tests,tests_extended}.rs
- connector.rs (698 LOC) → connector/{mod,operations,tests,tests_extended}.rs
- sqlite_backend.rs (693 LOC) → sqlite_backend/{mod,storage,search,tests}.rs

All public APIs preserved via re-exports. All 363 tests pass.
cargo clippy and cargo fmt clean.
